ported version 1.0.1 to singlePrecision
this includes the commits:
----------------------------------------------------------------------
From 7dab661bfc959e4258891591033b3ec536fe68d9 Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Tue, 28 Feb 2012 20:31:33 +0000
Subject: [PATCH 1/5] change: solveDynamicEq should now have better
 performance.

This function contained a loop in which a "current" and "next" set of
coordinates was calculated. This was redundant, as the "next" set of data
at iteration i is the "current" set at i+1.

----------------------------------------------------------------------
From 573aedb996bce6cf581a3ec8565af774e1340baa Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Thu, 1 Mar 2012 15:23:47 +0000
Subject: [PATCH 2/5] change: converted all functions from double precision to
 single precision

With the exception of those that are matlab-related as matlab formats use
double precision internally.

----------------------------------------------------------------------
From 7d7dadb117bbef031764a5f9bf16c24f7205a7f8 Mon Sep 17 00:00:00 2001
From: Emanuel Ey <emanuel.ey@gmail.com>
Date: Fri, 9 Mar 2012 15:57:31 +0000
Subject: [PATCH 3/5] change: created a test case to compare Single Precision
 to DP version.

Also added a tag to the version identifier to make it easy to recognize
distinguish between SP and DP.

+++ source/bracket.c
@@ -51,9 +51,9 @@
 #pragma  once
 #include <stdint.h>
 
-uintptr_t   bracket(uintptr_t, double*, double, uintptr_t*);
+uintptr_t   bracket(uintptr_t, float*, float, uintptr_t*);
 
-uintptr_t   bracket(uintptr_t n, double* x, double xi, uintptr_t* i){
+uintptr_t   bracket(uintptr_t n, float* x, float xi, uintptr_t* i){
     DEBUG(5, "Entering bracket().\n");
     uintptr_t   ia,im,ib;
