+++ source/linearSpaced.c
@@ -44,15 +44,15 @@
 
 #pragma once
 
-void    linearSpaced(uint32_t, double, double, double*);
+void    linearSpaced(uint32_t, float, float, float*);
 
-void    linearSpaced(uint32_t n, double xMin, double xMax, double* x){
+void    linearSpaced(uint32_t n, float xMin, float xMax, float* x){
     uintptr_t   i;
-    double  dx;
+    float  dx;
 
     DEBUG(10, "n:%u, xMin: %lf, xMax: %lf\n", n, xMin, xMax);
-    dx = (xMax - xMin)/((double)(n - 1));
+    dx = (xMax - xMin)/((float)(n - 1));
     for(i=0; i<n; i++){
-        x[i] = xMin + dx*(double)(i);
+        x[i] = xMin + dx*(float)(i);
     }
 }
